+++ finetuner/__init__.py
@@ -36,19 +36,25 @@ if TYPE_CHECKING:
 ft = Finetuner()
 
 
-def login(force: bool = False) -> None:
+def login(force: bool = False, interactive: Optional[bool] = None):
     """
     Login to Jina AI to use cloud-based fine-tuning. Thereby, an authentication token is
     generated which can be read with the :func:`~finetuner.get_token` function.
 
     :param force: If set to true, an existing token will be overwritten. Otherwise,
         you will not login again, if a valid token already exists.
+    :param interactive: Interactive mode should be set in Jupyter environments.
     """
-    ft.login(force=force)
+    ft.login(force=force, interactive=interactive)
 
 
 def notebook_login(force: bool = False):
-    ft.notebook_login(force=force)
+    warnings.warn(
+        message='Function `notebook_login` will be deprecated from Finetuner 0.7.0,'
+        'please use `login(interactive=True)` instead.',
+        category=DeprecationWarning,
+    )
+    ft.login(force=force, interactive=True)
 
 
 def list_callbacks() -> Dict[str, callback.CallbackStubType]:
@@ -124,6 +130,7 @@ def fit(
     num_workers: int = 4,
     to_onnx: bool = False,
     csv_options: Optional[CSVOptions] = None,
+    public: bool = False,
 ) -> Run:
     """Start a finetuner run!
 
@@ -188,6 +195,8 @@ def fit(
     :param csv_options: A :class:`CSVOptions` object containing options used for
         reading in training and evaluation data from a CSV file, if they are
         provided as such.
+    :param public: A boolean value indicates if the artifact is public. It should be
+        set to `True` if you would like to share your fine-tuned model with others.
 
     .. note::
        Unless necessary, please stick with `device="cuda"`, `cpu` training could be
@@ -219,6 +228,7 @@ def fit(
         num_workers=num_workers,
         to_onnx=to_onnx,
         csv_options=csv_options,
+        public=public,
     )